# PERF improve performance of is_lexsorted