--- a/pandas/_libs/algos.pyx
+++ b/pandas/_libs/algos.pyx
@@ -156,7 +156,6 @@ def is_lexsorted(list_of_arrays: list) -> bint:
         Py_ssize_t n, nlevels
         int64_t k, cur, pre
         ndarray arr
-        bint result = True
 
     nlevels = len(list_of_arrays)
     n = len(list_of_arrays[0])
@@ -181,9 +180,10 @@ def is_lexsorted(list_of_arrays: list) -> bint:
                     result = False
                     break
             if not result:
-                break
+                free(vecs)
+                return False
     free(vecs)
-    return result
+    return True
 
 
 @cython.boundscheck(False)